requirements.txt
PyYAML
flask
flask-swagger-ui
gunicorn
marshmallow
braintree
